allow reply_to in postdata to be [ null ]
sometimes the api will return [ null ] as the reply to field, to deal with this we need to allow the type to be null. @Tnix, does the meower API return [ null ] or is this a issue somewhere else in the process. It may be returning [] and then TS makes it [ null ].